Rollback. Removed code related to APIC Timer.
I ran into a problem when using PIT on PIC interrupts in VirtualBox.
After activating APIC, interrupts on vector 32 from PIT stopped being generated.
It is necessary to configure the APIC IO first.
On qemu and bochs this problem was not encountered, on them it was possible to use PIT in the old way, but not on Virtual Box.