UB HERE
There is a fucking UB going on somewhere and if the variable is not printed then the code fails, see ioapic.rs.
Looks like UB in code related to GeneralPurposeAllocator, I checked dlmalloc and acpi with miri and everything is ok there.
I don't know where UB comes from in the code, I need to check GeneralPurposeAllocator and related code.